add confirmation modal

+++ projects/tr-corporate/src/app/settings/permission/user-manage/user-manage.component.ts
@@ -29,6 +29,7 @@ import { merge, Observable, of as observableOf } from 'rxjs';
 import { catchError, map, startWith, switchMap } from 'rxjs/operators';
 import { getRoles } from '../../../utility/store/selectors/roles.selector';
 import { FilterService } from '../services/filter.service';
+import { ConfirmationComponent } from '../../../shared/components/confirmation/confirmation.component';
 
 @Component({
   selector: 'app-user-manage',
@@ -100,10 +101,6 @@ export class UserManageComponent implements OnInit {
     this._bottomSheet.open(MFilterComponent, { data: appliedFilterData }).afterDismissed()
       .subscribe(result => {
         if (result) {
-          // this.selectedRole = this.filterServ.SelectedRole;
-          // this.selectedStatus = this.filterServ.selectedStatus;
-          // this.sort.active = this.filterServ.selectedSort;
-
           this.selectedRole = result.filter_roletypeid;
           this.selectedStatus = result.filter_status;
           this.sort.active = result.sort;
@@ -208,16 +205,16 @@ export class UserManageComponent implements OnInit {
   }
 
   deleteUser(email: string) {
-    this.userServ.deleteUser({ 'email': email }).subscribe(res => {
-      if (res.error) {
-        // error from api
-        this.snackBar.open(res.message);
+    const dialogRef = this.dialog.open(ConfirmationComponent, { width: '500px', });
+
+    dialogRef.afterClosed().subscribe(isConfirmed => {
+      if (isConfirmed) {
+        this.userServ.deleteUser({ 'email': email }).subscribe(res => {
+          if (res.error) this.snackBar.open(res.message);
+          else this.snackBar.open(res.message);
+        });
       }
-      else {
-        // success from api
-        this.snackBar.open(res.message);
-      }
-    })
+    });
   }
 
   toggleUserActionMenu() {
